Control changes, updates for comp bot

+++ Blitzar/src/org/usfirst/frc/team4499/robot/commands/setReceiverPower.java
@@ -14,7 +14,7 @@ public class setReceiverPower extends Command {
 
 	private double setPower;
 	
-    public setReceiverPower(double power) {
+    public setReceiverPower(float power) {
         // Use requires() here to declare subsystem dependencies
         // eg. requires(chassis);
     	requires(Robot.receiver);
